requirements.txt
torch>=1.7.0
torchaudio>=0.7.0
asteroid_filterbanks==0.4.0
numpy
scipy
scikit-learn
tqdm
pyyaml
soundfile==0.11.0
pyannote.audio